Renames file
